Refactoring: - Moved cache collection to own cache to simplify manager - Renamed classes - Created package for cache - Simplified logging
